Cleanup the lineup section with our new format
