+++ src/_components/button/button.styles.css
@@ -9,4 +9,19 @@
     @apply bg-white
       text-black;
   }
+
+  &.theme-outline {
+    @apply bg-transparent
+      border
+      border-blue
+      duration-300
+      text-blue
+      /* text-gray */
+      transition-all;
+
+    &:hover {
+      @apply bg-blue
+        text-white;
+    }
+  }
 }
