--- src/_components/card/card.styles.css
@@ -1,3 +1,0 @@
-.component--card {
-  /* width: 25%; */
-}
