+++ src/_layouts/event.njk
@@ -44,49 +44,15 @@
     {% for speaker in speakers %}
       <div class="speaker grid gap-4 md:gap-8 max-w-4xl mb-8">
         <div class="text-center speaker--image mx-auto">
-          <span class="block mb-2">
-            {% image path=speaker.image, sm="256px", md="320px", lg="208px", xl="272px" %}
-          </span>
-          <strong class="block mb-2 text-lg">{{ speaker.title }}</strong>
-          <span class="block text-sm">
-            <em>{{ speaker.job_title }}</em>
-          </span>
-          <span class="block text-sm">
-            {{ speaker.company }}
-          </span>
-          {% if speaker.social.twitter or speaker.social.instagram or speaker.social.facebook %}
-            <div class="flex items-center justify-center mt-4">
-              {% if speaker.social.twitter %}
-                <a
-                  href="{{ speaker.social.twitter }}"
-                  target="_blank"
-                  class="mx-1 hover:opacity-75 transition-all duration-300"
-                >
-                  {% icon name="twitter" %}
-                </a>
-              {% endif %}
-              {% if speaker.social.facebook %}
-                <a
-                  href="{{ speaker.social.facebook }}"
-                  target="_blank"
-                  class="mx-1 hover:opacity-75 transition-all duration-300"
-                >
-                  {% icon name="facebook" %}
-                </a>
-              {% endif %}
-              {% if speaker.social.instagram %}
-                <a
-                  href="{{ speaker.social.instagram }}"
-                  target="_blank"
-                  class="mx-1 hover:opacity-75 transition-all duration-300"
-                >
-                  {% icon name="instagram" %}
-                </a>
-              {% endif %}
-            </div>
-          {% endif %}
+          {% speaker_card
+                image=speaker.image,
+                name=speaker.title,
+                job_title=speaker.job_title,
+                company=speaker.company,
+                social=speaker.social
+          %}
         </div>
-        <div class="">
+        <div>
           <h3>{{ speaker.talk_title }}</h3>
           <div class="mb-4">
             {% markdown %}{{ speaker.talk_body }}{% endmarkdown %}
