+++ src/events/theres-a-crack-in-the-pavement.md
@@ -10,6 +10,8 @@ body: |-
   Nulla laboris excepteur adipisicing non non amet occaecat. Ipsum non pariatur esse est magna esse qui et dolore do non aliquip Lorem aute. Consectetur minim proident quis deserunt ea id et labore laborum. Amet dolore magna in id ad. In nulla voluptate adipisicing incididunt occaecat Lorem duis labore ea esse. Ullamco incididunt sit sunt minim mollit.
 
   Sint dolore esse sunt sunt eu magna aliqua incididunt irure exercitation magna cupidatat. Laborum ullamco sit eu non id eu excepteur id do fugiat eiusmod. Quis culpa labore ea irure aute elit consequat deserunt labore ipsum. Sit amet aliqua elit sint excepteur ex deserunt exercitation.
+snippet: |-
+  Ex dolore consectetur cillum irure. Dolor nulla commodo nulla elit mollit Lorem sint. Duis exercitation consequat enim qui commodo ut. Nostrud id consequat excepteur eiusmod ea irure.
 layout: event
 speakers:
 - image: "/website/uploads/member-brandon.png"
